Add required GSI partition/sort key annotations
Add DynamoDbSecondaryPartitionKey and DynamoDbSecondarySortKey
annotations to the appropriate DynamoDB table data model to
enable the DynamoDB enhanced client to generate correct code
integrating with the ConsentsByServiceUser GSI.

The ConsentsByServiceUser GSI's compound key is the combination of
user ID and service ID, and we choose user ID as the GSI partition
key since it will be more well-distributed and hence introduce
fewer DynamoDB hot-key/throttling issues than service ID, which
we will take as the GSI sort key.

Before this change, querying the ListServiceUserConsents API
in AWS API Gateway results in a 500 Internal Service Error with
error message: "Attempt to execute an operation that requires a
secondary index without defining the index attributes in the table
metadata. Index name: ConsentsByServiceUser", and error type
java.lang.IllegalArgumentException.