Remove unit test dependency on active AWS credentials
GitHub workflows currently fail for this package when running
the ConsentManagementApiServiceTest.testDefaultConstructor()
due to lacking AWS credentials that the DynamoDB repository
construction implicitly depends on existing.

Extract logic for constructing the DynamoDB consent repository
to a helper that can be independently tested and removing the
unit test for the default service constructor which will not
be runnable from GitHub workflows.