Implement DDB mapper logic to conditionally set activeId
Update the logic for mapping Consent API models to the
ServiceUserConsent DynamoDB data model, to set activeId
to the consent partition key if the consent status is
ACTIVE and expiryTime is non-null, otherwise set
activeId to null.

This will be used to add/remove consents from the sparse
ActiveConsentsWithExpiryTime DynamoDB Global Secondary
Index (GSI), so that the GSI tracks all active consents
with expiry times.

This GSI will support the consent auto-expiry workflow.